Refactored OpenTelemetryTransactionProcessorTests
