+++ test/Sentry.Tests/ApiApprovalTests.Run.DotNet7_0.verified.txt
@@ -622,7 +622,7 @@ namespace Sentry
         public System.TimeSpan FlushTimeout { get; set; }
         public System.Net.IWebProxy? HttpProxy { get; set; }
         public System.TimeSpan InitCacheFlushTimeout { get; set; }
-        public Sentry.Instrumenter Instrumenter { get; set; }
+        public Sentry.Instrumenter Instrumenter { get; }
         public bool IsEnvironmentUser { get; set; }
         public bool IsGlobalModeEnabled { get; set; }
         public bool JsonPreserveReferences { get; set; }
